requirements.txt
streamlit==0.79.0
pandas==1.1.4
openpyxl==3.0.7
scikit-learn==0.24.1
sklearn==0.0
numpy==1.19.4
matplotlib==3.3.4
Pillow==8.0.1